Separate realms public endpoints
Use the `/public` prefix for the realms views that need to be public.
