+++ server/realms/backends/ldap/public_urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
-from . import views
+from . import public_views
 
 urlpatterns = [
     path('<uuid:uuid>/ldap/<uuid:session_pk>/login/',
-         views.LoginView.as_view(),
+         public_views.LoginView.as_view(),
          name="ldap_login"),
 ]
